cleanup: remove _test suffix
Renames harness and utility files (e.g., harness_test.go -> harness.go)
within the EPP integration test suite to align with project conventions.

Additionally moves shared global variables (k8sClient, testEnv) to
harness.go to ensure visibility for the renamed files, resolving
compiler scope issues that arise when separating test logic from
infrastructure code.
